+++ modules/alloy/src/main/ts/ephox/alloy/keying/FlowType.ts
@@ -70,13 +70,13 @@ const getKeydownRules = (
     KeyRules.rule(KeyMatch.inSet(westMovers), doMove(DomMovement.west(moveLeft, moveRight))),
     KeyRules.rule(KeyMatch.inSet(eastMovers), doMove(DomMovement.east(moveLeft, moveRight))),
     KeyRules.rule(KeyMatch.inSet(Keys.ENTER), execute),
-    KeyRules.rule(KeyMatch.inSet(Keys.SPACE), execute),
-    KeyRules.rule(KeyMatch.inSet(Keys.ESCAPE), doEscape)
+    KeyRules.rule(KeyMatch.inSet(Keys.SPACE), execute)
   ];
 };
 
 const getKeyupRules: () => Array<KeyRules.KeyRule<FlowConfig, Stateless>> = Fun.constant([
-  KeyRules.rule(KeyMatch.inSet(Keys.SPACE), KeyingTypes.stopEventForFirefox)
+  KeyRules.rule(KeyMatch.inSet(Keys.SPACE), KeyingTypes.stopEventForFirefox),
+  KeyRules.rule(KeyMatch.inSet(Keys.ESCAPE), doEscape)
 ]);
 
 export default KeyingType.typical(schema, NoState.init, getKeydownRules, getKeyupRules, () => Optional.some(focusIn));