+++ modules/tinymce/src/themes/silver/test/ts/headless/sketchers/SilverMenubarTest.ts
@@ -207,7 +207,7 @@ describe('headless.tinymce.themes.silver.sketchers.SilverMenubar Test', () => {
       await pWaitForMenuToAppear();
       await pAssertFocusOnToggleItem('Remember me');
       assertActiveToggleItemHasOneCheckmark('Remember me');
-      Keyboard.activeKeydown(doc, Keys.escape());
+      Keyboard.activeKeyup(doc, Keys.escape());
       await pAssertFocusOnMenuButton('Changes');
       await pWaitForMenuToDisappear();
 
@@ -233,10 +233,10 @@ describe('headless.tinymce.themes.silver.sketchers.SilverMenubar Test', () => {
       Keyboard.activeKeydown(doc, Keys.left());
       await pAssertFocusOnItem('Nested menu x 2');
 
-      Keyboard.activeKeydown(doc, Keys.escape());
+      Keyboard.activeKeyup(doc, Keys.escape());
       await pAssertFocusOnItem('Nested');
 
-      Keyboard.activeKeydown(doc, Keys.escape());
+      Keyboard.activeKeyup(doc, Keys.escape());
       await pAssertFocusOnMenuButton('Basic Menu Button');
       await pWaitForMenuToDisappear();
 
@@ -249,7 +249,7 @@ describe('headless.tinymce.themes.silver.sketchers.SilverMenubar Test', () => {
       Keyboard.activeKeydown(doc, Keys.enter());
       await pAssertFocusOnItem('Nested menu x 2');
 
-      Keyboard.activeKeydown(doc, Keys.escape());
+      Keyboard.activeKeyup(doc, Keys.escape());
       await pAssertFocusOnItem('Nested');
       Keyboard.activeKeydown(doc, Keys.up());
       Keyboard.activeKeydown(doc, Keys.enter());
@@ -262,7 +262,7 @@ describe('headless.tinymce.themes.silver.sketchers.SilverMenubar Test', () => {
       SilverMenubar.focus(getMenubar());
       await pAssertFocusOnMenuButton('Changes');
 
-      Keyboard.activeKeydown(doc, Keys.escape());
+      Keyboard.activeKeyup(doc, Keys.escape());
       store.assertEq('Pressing escape in menubar should fire event', [ 'Menubar.escape' ]);
     });